Add files via upload
In this assignment, you are required to answer the questions in the notebooks given
in the folders Part 1 and Part 2. The parts are independent from each other but it is
recommended to solve them in order. You may test your code using the test cases
provided to you. However, there are some hidden test cases as well. Please note that
your code must work for both types of tests to get full credit. Hardcoded values will
not be awarded any credit. Further details are given in the respective notebooks.