Add files via upload
In this assignment, you are required to apply your knowledge of machine learning models
to answer the questions in the notebooks given in the folders Part 1 and Part 2. The parts
are independent from each other but it is recommended to solve them in order. The
assignment will be checked both manually and through automated test cases via otter.